move css files to styles folder (they're not components!)
