+++ src/Components/ZipCodeForm.js
@@ -1,8 +1,7 @@
 import React, { useState } from 'react'
-
-import './form.css'
 import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
+import '../styles/form.css'
 
 const ZipCodeForm = ({ changeZip }) => {
