+++ src/styles/dailyDisplay.css
@@ -8,7 +8,7 @@
     padding-bottom: 12px;
 
 
-    flexWrap: 'wrap',
+    flex-wrap: 'wrap',
 }
 
 .time {
